requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
gdown
